Add extraction of Transaction type to common package
