+++ packages/common/src/types/index.ts
@@ -1,0 +1,7 @@
+export type Transaction = {
+  id?: string;
+  customer: string;
+  supplier: string;
+  total: string;
+  date: string;
+};
